% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bexis.get.dataset.R
\name{bexis.get.dataset_by}
\alias{bexis.get.dataset_by}
\title{Getting data from BExIS}
\usage{
bexis.get.dataset_by(id, base_url = bexis.options("base_url"))
}
\arguments{
\item{id}{Is the ID of the dataset you want to access}

\item{base_url}{Is the URL that is use to talk to}
}
\description{
The function provides access to dataset in BExIS.
}

