% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bexis.datasets.get.R
\name{bexis.get.structures}
\alias{bexis.get.structures}
\title{Getting all structures from BExIS}
\usage{
bexis.get.structures()
}
\arguments{
\item{base_url}{Is the URL of the BExIS instance (e.g. www.mybexis.de)
If the base url is already provided by the function bexis.options()
 then this value will be taken and you do not need to provide it again.}

\item{token}{is an authorization key to access the data in Bexis. Get the token key in practice user profile.}
}
\value{
The function returns a list of structure ids.
An error is provided in case the data is not found or if you don't
have the rights to access it.
}
\description{
The function provides access to structure that on a BExIS II instance.
}
\examples{
\dontrun{
        bexis.get.structures()
      }
}
