% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bexis.datasets.get.R
\name{bexis.GetDatasetIds}
\alias{bexis.GetDatasetIds}
\title{Getting a single metadata structure from BExIS}
\usage{
bexis.GetDatasetIds(id)
}
\description{
The function provides access to data that on a BExIS II instance.
An error is provided in case the data is not found or if you don't
have the rights to access it.
}
\examples{
\dontrun{
        bexis.GetDatasetIds()
      }
}
