% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bexis.metadata.get.R
\name{bexis.GetDataStructures}
\alias{bexis.GetDataStructures}
\title{Getting data structures from BExIS}
\usage{
bexis.GetDataStructures()
}
\value{
The function returns a metadata of the requested dataset.
An error is provided in case the data is not found or if you don't
have the rights to access it.
}
\description{
The function provides access to data that on a BExIS II instance.
}
\examples{
\dontrun{
        bexis.GetDataStructures()
      }
}
