% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bexis.metadata.get.R
\name{bexis.GetMetadataById}
\alias{bexis.GetMetadataById}
\title{Getting metadata from BExIS}
\usage{
bexis.GetMetadataById(id)
}
\arguments{
\item{id}{Is the ID of the dataset you want to access}
}
\value{
The function returns a metadata of the requested dataset.
An error is provided in case the data is not found or if you don't
have the rights to access it.
}
\description{
The function provides access to data that on a BExIS II instance.
}
\examples{
\dontrun{
        bexis.GetMetadataById(8)
      }
}
