% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bexis.options.R
\name{bexis.options}
\alias{bexis.options}
\title{Set or query otions used for the BExIS R package.}
\usage{
bexis.options(...)
}
\arguments{
\item{\dots}{unbound parameters}
}
\description{
This function is used to query and set rBExIS R package specific options
like the URL to the BEFdata portal and the URL to the tematres thesaurus.
}

