% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bexis.helper.R
\name{get_download_url}
\alias{get_download_url}
\title{Functions to provide helpers for other functions
e.g. paste together a url after certain criteria}
\usage{
get_download_url(base_url, version)
}
\description{
Functions to provide helpers for other functions
e.g. paste together a url after certain criteria
}

