% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bexis.metadata.get.R
\name{GetMetadata}
\alias{GetMetadata}
\title{Getting all metadata objects from BExIS}
\usage{
GetMetadata()
}
\value{
The function returns a list of metadata objects.
An error is provided in case the data is not found or if you don't
have the rights to access it.
}
\description{
The function provides access to data that on a BExIS II instance.
}
\examples{
\dontrun{
        GetMetadata()
      }
}
