% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bexis.get.R
\name{bexis.get.dataset_by}
\alias{bexis.get.dataset_by}
\title{Getting data from BExIS}
\usage{
bexis.get.dataset_by(id, base_url = bexis.options("base_url"),
  token = bexis.options("token"))
}
\arguments{
\item{id}{Is the ID of the dataset you want to access}

\item{base_url}{Is the URL of the BExIS instance (e.g. www.mybexis.de)
If the base url is already provided by the function bexis.options()
 then this value will be taken and you do not need to provide it again.}

\item{token}{is an authorization key to access the data in Bexis. Get the token key in practice user profile.}
}
\value{
The function returns a data frame of the requested dataset.
An error is provided in case the data is not found or if you don't
have the rights to access it.
}
\description{
The function provides access to data that on a BExIS II instance.
}
\examples{
\dontrun{
        bexis.get.dataset_by(id=8)
      }
}
